requirements/requirements_train.txt
numpy==1.23.5
smplx==0.1.28
hydra-core==1.3.2
easydict
tqdm
open3d
lxml
ray
ipdb
joblib
scipy
jupyter
loguru
tensorboard
mujoco
mink
dm_control
loop_rate_limiters
qpsolvers[quadprog]
accelerate
tabulate
matplotlib
pandas
lmdb
termcolor
rich
pytorch-tcn
einops
onnxruntime-gpu
protobuf==3.20.3
onnx
pre-commit
ruff
pytest
imageio>=2.9
imageio-ffmpeg

# human_body_prior 依赖
human-body-prior